backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
brotli==1.2.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
distro==1.9.0
Django==5.2.5
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
fastapi==0.136.1
filelock==3.29.0
fsspec==2026.3.0
gradio==6.13.0
gradio_client==2.5.0
groovy==0.1.2
groq==1.2.0
gunicorn==25.3.0
h11==0.16.0
hf-gradio==0.4.1
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.12.0
idna==3.13
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
orjson==3.11.8
packaging==26.2
pandas==3.0.2
pillow==12.2.0
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
pydub==0.25.1
Pygments==2.20.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
sentence-transformers==5.4.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomlkit==0.14.0
torch==2.11.0
tqdm==4.67.3
transformers==5.6.2
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
